% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pre_and_post_lsff_summaries_afe.R
\name{calculate_pre_and_post_lsff_summaries_afe}
\alias{calculate_pre_and_post_lsff_summaries_afe}
\title{Calculate Pre and Post LSFF Nutrient Summaries}
\usage{
calculate_pre_and_post_lsff_summaries_afe(
  householdConsumptionDf = householdConsumption,
  householdDetailsDf = householdDetails,
  nctListDf = nctList,
  intakeThresholdsDf = intakeThresholds,
  aggregationGroup = c("admin0Name", "admin1Name"),
  fortifiableFoodItemsDf = createFortifiableFoodItemsTable(),
  foodVehicleName = "wheat flour",
  fortificationLevelsDf = fortificationLevels,
  years = c(2021:2024),
  MNList = "A"
)
}
\arguments{
\item{householdConsumptionDf}{A dataframe containing household consumption data. Must contain columns: "householdId", "amountConsumedInG".}

\item{householdDetailsDf}{A dataframe containing household details. Must contain columns: "householdId", "memberCount","afeFactor".}

\item{nctListDf}{A dataframe containing nutrient composition tables. Must contain column: "nutrient".}

\item{intakeThresholdsDf}{A dataframe containing intake thresholds for nutrients. Must contain columns: "nutrient", "ear", "ul".}

\item{aggregationGroup}{A character vector of administrative groups to aggregate the data. Must not be empty. Defaults to c("admin0Name", "admin1Name").}

\item{fortifiableFoodItemsDf}{A dataframe containing fortifiable food items. Generated using the function \code{createFortifiableFoodItemsTable()}.}

\item{foodVehicleName}{A character string specifying the name of the food vehicle for fortification. Defaults to "wheat flour".}

\item{fortificationLevelsDf}{A dataframe containing the average fortification levels for different nutrients and years.}

\item{years}{A numeric vector specifying the years for which LSFF is analyzed. Defaults to 2021:2024.}

\item{MNList}{A character vector of nutrients to be included in the analysis. Defaults to "A". Must not be empty.}
}
\value{
A dataframe with the summaries of nutrient inadequacy for the specified administrative groups before and after LSFF.
}
\description{
This function calculates summaries of nutrient inadequacy before and after large-scale food fortification (LSFF) for different administrative groups using the Adult Female Equivalent (AFE) Method.
}
\examples{
\dontrun{
calculate_pre_and_post_lsff_summaries_afe(
    householdConsumptionDf = householdConsumption,
    householdDetailsDf = householdDetails,
    nctListDf = nctList,
    intakeThresholdsDf = intakeThresholds,
    aggregationGroup = c("admin0Name", "admin1Name"),
    fortifiableFoodItemsDf = createFortifiableFoodItemsTable(),
    foodVehicleName = "wheat flour",
    fortificationLevels = fortificationLevels,
    years = c(2021:2024),
    MNList = c("A", "Ca")
)
}
}
