% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculateBaselineInadequacyAfe}
\alias{calculateBaselineInadequacyAfe}
\title{Calculate Baseline Nutrient Inadequacy}
\usage{
calculateBaselineInadequacyAfe(
  householdConsumptionDf = householdConsumption,
  householdDetailsDf = householdDetails,
  nctListDf = nctList,
  intakeThresholdsDf = intakeThresholds,
  aggregationGroup = c("admin0Name", "admin1Name"),
  MNList = c("Ca", "Carbohydrates", "Cu", "Energy", "Fat", "Fe", "Fibre", "I", "IP6",
    "Mg", "Protein", "Se", "Zn", "Ash", "B6", "B2", "D", "N", "K", "P", "Moisture",
    "Cholesterol", "E", "Na", "A", "C", "B12", "B1", "B3", "B9", "B5", "B7", "Mn")
)
}
\arguments{
\item{aggregationGroup}{A character vector of administrative groups. Must not be empty.}

\item{MNList}{A character vector of nutrients. If empty, defaults to a list of all nutrients.}

\item{dataDir}{The directory where the data is stored.}
}
\value{
A dataframe with the baseline inadequacy of nutrients for the specified administrative groups.
}
\description{
This function calculates the baseline inadequacy of nutrients for different administrative groups.
}
\examples{
calculateBaselineInadequacyAfe(MNList = c("A", "Ca"))
}
