% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculateBaselineInadequacyCND}
\alias{calculateBaselineInadequacyCND}
\title{Calculate Baseline Nutrient Inadequacy}
\usage{
calculateBaselineInadequacyCND(
  MNList = c("A"),
  aggregationGroup = c("admin0Name", "admin1Name"),
  dataDir = here::here("data/sd123")
)
}
\arguments{
\item{MNList}{A character vector of nutrients. If empty, defaults to a list of all nutrients.}

\item{aggregationGroup}{A character vector of administrative groups. Must not be empty.}

\item{dataDir}{The directory where the data is stored.}
}
\value{
A dataframe with the baseline inadequacy of nutrients for the specified administrative groups.
}
\description{
This function calculates the baseline inadequacy of nutrients for different administrative groups.
}
\examples{
calculateBaselineInadequacy(MNList = c("A", "Ca"))
}
