% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateBaselinePrevalenceHighIntakeRiskAfe.R
\name{calculateBaselinePrevalenceHighIntakeRiskAfe}
\alias{calculateBaselinePrevalenceHighIntakeRiskAfe}
\title{Calculate Baseline High Intake Risk of Nutrients (AFE Method)}
\usage{
calculateBaselinePrevalenceHighIntakeRiskAfe(
  householdConsumptionDf = householdConsumption,
  householdDetailsDf = householdDetails,
  nctListDf = nctList,
  intakeThresholdsDf = intakeThresholds,
  aggregationGroup = c("admin0Name", "admin1Name"),
  MNList = c("Ca", "Carbohydrates", "Cu", "Energy", "Fat", "Fe", "Fibre", "I", "IP6",
    "Mg", "Protein", "Se", "Zn", "Ash", "B6", "B2", "D", "N", "K", "P", "Moisture",
    "Cholesterol", "E", "Na", "A", "C", "B12", "B1", "B3", "B9", "B5", "B7", "Mn")
)
}
\arguments{
\item{householdConsumptionDf}{A dataframe containing household consumption data. Must contain columns: "householdId", "amountConsumedInG", "memberCount".}

\item{householdDetailsDf}{A dataframe containing household details. Must contain column: "householdId".}

\item{nctListDf}{A dataframe containing nutrient composition tables. Must contain columns: "nutrient", "foodId".}

\item{intakeThresholdsDf}{A dataframe containing intake thresholds for nutrients. Must contain columns: "nutrient", "CND".}

\item{aggregationGroup}{A character vector of administrative groups to aggregate the data. Must not be empty. Defaults to c("admin0Name", "admin1Name").}

\item{MNList}{A character vector of nutrients to be included in the analysis. If empty, defaults to a comprehensive list of nutrients.}
}
\value{
A dataframe with the baseline inadequacy of nutrients for the specified administrative groups.
}
\description{
This function calculates the baseline risk of high nutrient intake for different administrative groups using the Adequate Food Energy (AFE) Method. It compares daily apparent micronutrient intake per AFE to the tolerable upper intake levels (UL) to classify household diets as at risk of high intake. The results are summarized at national and subnational levels.
}
\examples{
calculateBaselineInadequacyAfe(
    householdConsumptionDf = householdConsumption,
    householdDetailsDf = householdDetails,
    nctListDf = nctList,
    intakeThresholdsDf = intakeThresholds,
    aggregationGroup = c("admin0Name", "admin1Name"),
    MNList = c("Ca", "Carbohydrates")
)
}
