% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculateConsPerAfe}
\alias{calculateConsPerAfe}
\title{Calculate Consumption Per Adult Female Equivalent (AFE)}
\usage{
calculateConsPerAfe(df, consumptionCol, afeCol)
}
\arguments{
\item{df}{A dataframe containing food consumption data and AFE data.}

\item{consumptionCol}{A string specifying the column name for the food consumption data.}

\item{afeCol}{A string specifying the column name for the AFE data.}
}
\value{
A dataframe with the calculated consumption per AFE for each food item.
}
\description{
This function calculates the consumption per AFE for each food item in a dataframe.
}
\examples{
\dontrun{
df <- data.frame(food = c("apple", "banana", "carrot"), consumption = c(100, 200, 150), afe = c(2, 3, 4))
df <- calculateConsPerAfe(df, "consumption", "afe")
}

}
