% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculateFoodVehicleHouseholdReach}
\alias{calculateFoodVehicleHouseholdReach}
\title{Calculate Food Vehicle Household Reach}
\usage{
calculateFoodVehicleHouseholdReach(
  householdConsumptionDf = householdConsumption,
  householdDetailsDf = householdDetails,
  fortifiableFoodItemsDf = fortifiable_food_items,
  foodVehicleName = "wheat flour",
  aggregationGroup = c("admin0Name", "admin1Name")
)
}
\arguments{
\item{householdConsumptionDf}{A dataframe containing household consumption data. Must contain columns: "foodGenusId", "householdId".}

\item{householdDetailsDf}{A dataframe containing household details. Must contain column: "householdId".}

\item{fortifiableFoodItemsDf}{A dataframe containing fortifiable food items. Must contain columns: "food_genus_id", "food_vehicle_name".}

\item{foodVehicleName}{A string specifying the food vehicle name. Default is "wheat flour".}

\item{aggregationGroup}{A character vector specifying the columns to group by. Default is \code{c("admin0Name", "admin1Name")}.}
}
\value{
A dataframe with the household reach for the specified food vehicle.
}
\description{
This function calculates the household reach for each food vehicle. It filters the fortifiable food items to get the food vehicle, processes the consumption data, and then calculates the household reach.
}
\examples{
calculateFoodVehicleHouseholdReach(
    householdConsumptionDf = householdConsumption,
    householdDetailsDf = householdDetails,
    fortifiableFoodItemsDf = fortifiable_food_items,
    foodVehicleName = "wheat flour",
    aggregationGroup = c("admin0Name", "admin1Name")
)

}
