% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculateMNsContent}
\alias{calculateMNsContent}
\title{Calculate Micronutrient Content}
\usage{
calculateMNsContent(df, consumptionCol, MNs)
}
\arguments{
\item{df}{A dataframe containing food consumption data and micronutrient content.}

\item{consumptionCol}{A string specifying the column name for the food consumption data.}

\item{MNs}{A vector of column names for the micronutrients.}
}
\value{
A dataframe with the calculated micronutrient content for each food item.
}
\description{
This function calculates the micronutrient content for each food item in a dataframe.
}
\examples{
\dontrun{
df <- data.frame(
    food = c("apple", "banana", "carrot"),
    consumption = c(100, 200, 150), vitaminA = c(0.5, 0.6, 0.7)
)
df <- calculateMNsContent(df, "consumption", c("vitaminA"))
}

}
