% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkData.R
\name{checkData}
\alias{checkData}
\title{Check if Required Data is Loaded}
\usage{
checkData()
}
\value{
A logical value indicating whether the required data is loaded. Returns TRUE if all required data is loaded, and FALSE otherwise.
}
\description{
This function checks if the required data ("householdConsumption", "householdDetails", and "nctList") is loaded in the global environment.
}
\examples{
\dontrun{
checkData()
}

}
\keyword{internal}
