% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMasterNct.R
\name{createMasterNct}
\alias{createMasterNct}
\title{Create a master NCT (Nutrient Composition Table)}
\usage{
createMasterNct(
  nctList,
  fctListIdCol = "fctListId",
  foodGenusIdCol = "foodGenusId",
  micronutrientIdCol = "micronutrientId",
  micronutrientCompositionCol = "micronutrientComposition"
)
}
\arguments{
\item{nctList}{A list of NCTs.}

\item{fctListIdCol}{A string specifying the column name for the FCT list ID. Default is "fctListId".}

\item{foodGenusIdCol}{A string specifying the column name for the food genus ID. Default is "foodGenusId".}

\item{micronutrientIdCol}{A string specifying the column name for the micronutrient ID. Default is "micronutrientId".}

\item{micronutrientCompositionCol}{A string specifying the column name for the micronutrient composition. Default is "micronutrientComposition".}
}
\value{
A data frame representing the master NCT. Each row corresponds to a food item, and each column corresponds to a micronutrient.
}
\description{
This function creates a master NCT from a list of NCTs from the MAPS Tool. It selects specified columns from the NCT list and reshapes the data from long to wide format.
}
\examples{
\dontrun{
createMasterNct(nctList)
}

}
\keyword{internal}
