% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{enforceNumeric}
\alias{enforceNumeric}
\title{Enforce Numeric Data Type for Specified Columns}
\usage{
enforceNumeric(df, cols)
}
\arguments{
\item{df}{A dataframe.}

\item{cols}{A vector of column names that should be converted to numeric.}
}
\value{
A dataframe with the specified columns converted to numeric.
}
\description{
This function converts specified columns of a dataframe to numeric data type.
}
\examples{
\dontrun{
df <- data.frame(a = c("1", "2", "3"), b = c("4", "5", "6"))
df <- enforceNumeric(df, c("a", "b"))
}

}
