% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMnThresholds.R
\name{getMnThresholds}
\alias{getMnThresholds}
\title{Get Micronutrient Thresholds}
\usage{
getMnThresholds(intakeThresholds, nutrient, param = "ear")
}
\arguments{
\item{intakeThresholds}{A dataframe containing intake thresholds for various micronutrients.}

\item{param}{A string specifying the parameter to retrieve. Can be "ear", "ul", or "unitAdequacy". Default is "ear".}

\item{Mn}{A string specifying the micronutrient to retrieve thresholds for.}
}
\value{
A list containing the thresholds for the specified micronutrient. If param is "ul" or "unitAdequacy", a single value is returned.
}
\description{
This function retrieves the thresholds for a specified micronutrient from a dataframe of intake thresholds.
}
\examples{
\dontrun{
intakeThresholds <- data.frame(nutrient = c("vitaminA", "vitaminB", "vitaminC"),
 ear = c(0.5, 0.6, 0.7), ul = c(1, 1.2, 1.3), unitAdequacy = c(0.8, 0.9, 1))
thresholds <- getMnThresholds(intakeThresholds, "vitaminA", "ul")
}

}
\keyword{internal}
