% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMnThresholdUnits.R
\name{getMnThresholdUnits}
\alias{getMnThresholdUnits}
\title{Get Micronutrient Thresholds Units}
\usage{
getMnThresholdUnits(intakeThresholds, nutrient, param = "unitAdequacy")
}
\arguments{
\item{intakeThresholds}{A dataframe containing intake thresholds for various micronutrients.}

\item{param}{A string specifying the parameter to retrieve.Default is "unitAdequacy".}

\item{Mn}{A string specifying the micronutrient to retrieve thresholds for.}
}
\value{
A list containing the thresholds units for the specified micronutrient.
}
\description{
This function retrieves the thresholds for a specified micronutrient from a dataframe of intake thresholds.
}
\examples{
\dontrun{
intakeThresholds <- data.frame(
    nutrient = c("vitaminA", "vitaminB", "vitaminC"),
    ear = c(0.5, 0.6, 0.7), ul = c(1, 1.2, 1.3), unitAdequacy = c(0.8, 0.9, 1)
)
thresholds <- getMnThresholdUnits(intakeThresholds, "vitaminA", "unitAdequacy")
}

}
\keyword{internal}
