% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rda_files.R
\name{load_rda_files}
\alias{load_rda_files}
\title{Load RDA Files from a Directory}
\usage{
load_rda_files(dir = "data/sd123")
}
\arguments{
\item{dir}{A character string specifying the directory from which to load the \code{.rda} files. Defaults to \code{"data/sd123"}.}
}
\value{
This function does not return a value. It loads the \code{.rda} files into the global environment.
}
\description{
This function lists all \code{.rda} files in a specified directory and loads them into the global environment.
}
\examples{
\dontrun{
# Load .rda files from the default directory
load_rda_files()

# Load .rda files from a custom directory
load_rda_files("your/custom/path")
}

}
\keyword{internal}
