% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{loadMapsRdaTables}
\alias{loadMapsRdaTables}
\title{Load all .rda files in a directory}
\usage{
loadMapsRdaTables(dir)
}
\arguments{
\item{dir}{A character string specifying the directory to load .rda files from.}
}
\value{
Invisible NULL. This function is called for its side effect of loading .rda files into the global environment.
}
\description{
This function loads all .rda files in the specified directory into the global environment.
}
\examples{
\dontrun{
load_all_rda_files("data/sd123")
}

}
