% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{previewData}
\alias{previewData}
\title{Preview the First 1000 Rows of a Dataframe}
\usage{
previewData(df)
}
\arguments{
\item{df}{A dataframe to preview.}
}
\value{
A DT::datatable object showing the first 1000 rows of the input dataframe.
}
\description{
This function previews the first 1000 rows of a dataframe using DT::datatable.
The name of the dataframe is used as the caption, which is left-aligned.
}
\examples{
# Create a dataframe
df <- data.frame(
    x = rnorm(5000),
    y = rnorm(5000),
    z = rnorm(5000)
)
# Preview the dataframe
previewData(df)

}
