% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAverageFortificationLevels.R
\name{yearAverageFortificationLevel}
\alias{yearAverageFortificationLevel}
\title{Year Average Fortification Level}
\usage{
yearAverageFortificationLevel(
  fortification_vehicle,
  Year,
  MN,
  fortificationLevels =
    readr::read_csv(here::here("data/average_fortification_levels.csv"))
)
}
\arguments{
\item{fortification_vehicle}{A string specifying the fortification vehicle.}

\item{Year}{An integer specifying the year.}

\item{MN}{A string specifying the micronutrient.}

\item{fortificationLevels}{A dataframe containing the average fortification levels.}
}
\value{
A numeric value representing the average fortification level for the specified year and micronutrient.
}
\description{
This function loads the average fortification levels from a CSV file located in the data directory,
and then returns the average fortification level for a specific year and micronutrient.
}
\examples{
\dontrun{
avgFortificationLevel <- yearAverageFortificationLevel("wheat flour", 2024, "A")
}
}
\keyword{internal}
